feat: 추가 가능한 input 래퍼 공통 컴포넌트로 만들기
